+++ src/partnermax/types/__init__.py
@@ -8,6 +8,4 @@ from .key_issue_params import KeyIssueParams as KeyIssueParams
 from .key_list_response import KeyListResponse as KeyListResponse
 from .dealer_list_params import DealerListParams as DealerListParams
 from .key_issue_response import KeyIssueResponse as KeyIssueResponse
-from .dealer_create_params import DealerCreateParams as DealerCreateParams
-from .dealer_list_response import DealerListResponse as DealerListResponse
 from .dealer_update_params import DealerUpdateParams as DealerUpdateParams