+++ tests/api_resources/test_dealers.py
@@ -9,93 +9,14 @@ import pytest
 
 from partnermax import Partnermax, AsyncPartnermax
 from tests.utils import assert_matches_type
-from partnermax.types import (
-    DealerDetail,
-    DealerListResponse,
-)
+from partnermax.types import DealerDetail, DealerSummary
+from partnermax.pagination import SyncCursorPage, AsyncCursorPage
 
 base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 
 
 class TestDealers:
     parametrize = pytest.mark.parametrize("client", [False, True], indirect=True, ids=["loose", "strict"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    def test_method_create(self, client: Partnermax) -> None:
-        dealer = client.dealers.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        )
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    def test_method_create_with_all_params(self, client: Partnermax) -> None:
-        dealer = client.dealers.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-            activate=True,
-            metadata={"foo": "string"},
-            idempotency_key="Idempotency-Key",
-        )
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    def test_raw_response_create(self, client: Partnermax) -> None:
-        response = client.dealers.with_raw_response.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        )
-
-        assert response.is_closed is True
-        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
-        dealer = response.parse()
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    def test_streaming_response_create(self, client: Partnermax) -> None:
-        with client.dealers.with_streaming_response.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        ) as response:
-            assert not response.is_closed
-            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
-
-            dealer = response.parse()
-            assert_matches_type(DealerDetail, dealer, path=["response"])
-
-        assert cast(Any, response.is_closed) is True
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -203,7 +124,7 @@ class TestDealers:
     @parametrize
     def test_method_list(self, client: Partnermax) -> None:
         dealer = client.dealers.list()
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(SyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -213,7 +134,7 @@ class TestDealers:
             limit=1,
             status="active",
         )
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(SyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -223,7 +144,7 @@ class TestDealers:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         dealer = response.parse()
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(SyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -233,7 +154,7 @@ class TestDealers:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             dealer = response.parse()
-            assert_matches_type(DealerListResponse, dealer, path=["response"])
+            assert_matches_type(SyncCursorPage[DealerSummary], dealer, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -284,83 +205,6 @@ class TestAsyncDealers:
     parametrize = pytest.mark.parametrize(
         "async_client", [False, True, {"http_client": "aiohttp"}], indirect=True, ids=["loose", "strict", "aiohttp"]
     )
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    async def test_method_create(self, async_client: AsyncPartnermax) -> None:
-        dealer = await async_client.dealers.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        )
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    async def test_method_create_with_all_params(self, async_client: AsyncPartnermax) -> None:
-        dealer = await async_client.dealers.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-            activate=True,
-            metadata={"foo": "string"},
-            idempotency_key="Idempotency-Key",
-        )
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    async def test_raw_response_create(self, async_client: AsyncPartnermax) -> None:
-        response = await async_client.dealers.with_raw_response.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        )
-
-        assert response.is_closed is True
-        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
-        dealer = await response.parse()
-        assert_matches_type(DealerDetail, dealer, path=["response"])
-
-    @pytest.mark.skip(reason="Mock server tests are disabled")
-    @parametrize
-    async def test_streaming_response_create(self, async_client: AsyncPartnermax) -> None:
-        async with async_client.dealers.with_streaming_response.create(
-            address="xx",
-            business_name="xx",
-            city="xx",
-            contact_email="dev@stainless.com",
-            contact_phone="xxxxx",
-            postal_code="21029",
-            primary_domain="29-0.mi-57.16u-2d91-aha.o-l7c19m0.z.9zhin-8ja2-7.447----86.6.61--5-6.2.3-i2al.r.name",
-            province_code="SE",
-            vat_number="IT21029798095",
-        ) as response:
-            assert not response.is_closed
-            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
-
-            dealer = await response.parse()
-            assert_matches_type(DealerDetail, dealer, path=["response"])
-
-        assert cast(Any, response.is_closed) is True
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -468,7 +312,7 @@ class TestAsyncDealers:
     @parametrize
     async def test_method_list(self, async_client: AsyncPartnermax) -> None:
         dealer = await async_client.dealers.list()
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(AsyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -478,7 +322,7 @@ class TestAsyncDealers:
             limit=1,
             status="active",
         )
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(AsyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -488,7 +332,7 @@ class TestAsyncDealers:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         dealer = await response.parse()
-        assert_matches_type(DealerListResponse, dealer, path=["response"])
+        assert_matches_type(AsyncCursorPage[DealerSummary], dealer, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -498,7 +342,7 @@ class TestAsyncDealers:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             dealer = await response.parse()
-            assert_matches_type(DealerListResponse, dealer, path=["response"])
+            assert_matches_type(AsyncCursorPage[DealerSummary], dealer, path=["response"])
 
         assert cast(Any, response.is_closed) is True